Revert "exceptions - move to root folder instead of under converters"
This reverts commit 933284005be42b5db5af2a0992ef4e137d5e282e.
