+++ src/convertool/converters/converter_msg.py
@@ -18,7 +18,6 @@ from extract_msg.msg_classes import MessageBase
 from extract_msg.msg_classes import MessageSigned
 from striprtf.striprtf import rtf_to_text
 
-from convertool.exceptions import ConvertError
 from convertool.util import TempDir
 
 from .base import _shared_dependencies
@@ -28,6 +27,7 @@ from .base import ConverterABC
 from .base import dummy_base_file
 from .converter_html import ConverterHTML
 from .converter_pdf import ConverterPDFToImage
+from .exceptions import ConvertError
 
 
 # noinspection DuplicatedCode
