+++ src/convertool/converters/converter_pdf.py
@@ -1,7 +1,6 @@
 from pathlib import Path
 from typing import ClassVar
 
-from convertool.exceptions import ConvertError
 from convertool.util import TempDir
 
 from .base import _shared_dependencies
@@ -9,6 +8,7 @@ from .base import _shared_platforms
 from .base import _shared_process_timeout
 from .base import ConverterABC
 from .converter_image import ConverterImage
+from .exceptions import ConvertError
 
 
 class ConverterPDF(ConverterABC):
