+++ src/convertool/converters/converter_tnef.py
@@ -13,9 +13,8 @@ from tnefparse import TNEF
 from tnefparse import TNEFObject
 from tnefparse.mapi import TNEFMAPI_Attribute
 
-from convertool.exceptions import ConvertError
-
 from .base import ConverterABC
+from .exceptions import ConvertError
 
 
 @dataclass
