+++ src/convertool/converters/exceptions.py
